requirements.txt
PyYAML==6.0
py-markdown-table==1.1.0
requests==2.32.3
